Refactor setting components into their own file
